+++ test/lib/expectations.js
@@ -2,21 +2,6 @@ const chai = require('chai');
 
 const expect = chai.expect;
 
-function findHelpPageBase($) {
-  expect($('.local-header--title--question').text().trim()).to.match(/^Find a pharmacy/);
-  expect($('#location').is('input')).is.equal(true);
-}
-
-function findHelpPageInvalidEntry($) {
-  findHelpPageBase($);
-  expect($('label[for=location]').text()).to.contain('Enter a valid postcode');
-}
-
-function findHelpPage($) {
-  findHelpPageBase($);
-  expect($('label[for=location]').text()).to.contain('Enter a town, city or postcode');
-}
-
 function htmlWith200Status(res) {
   expect(res).to.have.status(200);
   expect(res).to.be.html;
@@ -72,8 +57,6 @@ function searchAgainPage($) {
 module.exports = {
   call111Callout,
   disambiguationPageBreadcrumb,
-  findHelpPage,
-  findHelpPageInvalidEntry,
   htmlWith200Status,
   midsomerNortonResults,
   noResultsPageBreadcrumb,