+++ test/unit/lib/pageTitles.js
@@ -1,4 +1,5 @@
 const chai = require('chai');
+const appTitle = require('../../../app/lib/constants').app.title;
 const pageTitles = require('../../../app/lib/pageTitles');
 
 const expect = chai.expect;
@@ -9,21 +10,21 @@ describe('displayTitles', () => {
       const location = '';
       const errorMessage = undefined;
       const result = pageTitles.search(location, errorMessage);
-      expect(result).to.equal('Find a pharmacy');
+      expect(result).to.equal(appTitle);
     });
 
     it('should return \'Find a pharmacy - Enter a town, city or postcode\' when error message, and no location provided', () => {
       const location = '';
       const errorMessage = 'Please enter a location';
       const result = pageTitles.search(location, errorMessage);
-      expect(result).to.equal('Find a pharmacy - Enter a town, city or postcode, or use your location');
+      expect(result).to.equal(`${appTitle} - Enter a town, city or postcode, or use your location`);
     });
 
     it('should return \'Find a pharmacy - We can\'t find the postcode ...\' with uppercase location when error message, and location provided', () => {
       const location = 'notapostcode';
       const errorMessage = 'Please enter a location';
       const result = pageTitles.search(location, errorMessage);
-      expect(result).to.equal('Find a pharmacy - We can\'t find the postcode \'NOTAPOSTCODE\'');
+      expect(result).to.equal(`${appTitle} - We can't find the postcode 'NOTAPOSTCODE'`);
     });
   });
 
@@ -32,14 +33,14 @@ describe('displayTitles', () => {
       const location = 'no such place';
       const places = [];
       const result = pageTitles.disambiguation(location, places);
-      expect(result).to.equal('Find a pharmacy - We can\'t find \'no such place\'');
+      expect(result).to.equal(`${appTitle} - We can't find 'no such place'`);
     });
 
     it('should return \'Find a pharmacy - places that match ...\' when places provided', () => {
       const location = 'Leeds';
       const places = [{ name: 'Leeds1' }, { name: 'Leeds2' }];
       const result = pageTitles.disambiguation(location, places);
-      expect(result).to.equal('Find a pharmacy - Places that match \'Leeds\'');
+      expect(result).to.equal(`${appTitle} - Places that match 'Leeds'`);
     });
   });
 
@@ -48,7 +49,7 @@ describe('displayTitles', () => {
       const location = 'Remote location';
       const services = [];
       const result = pageTitles.results(location, services);
-      expect(result).to.equal('Find a pharmacy - We can\'t find any pharmacies near Remote location');
+      expect(result).to.equal(`${appTitle} - We can't find any pharmacies near Remote location`);
     });
 
     it('should return \'Find a pharmacy - Pharmacies near ...\' with friendly location when services are available', () => {