+++ app/lib/canonicalUrl.js
@@ -1,0 +1,1 @@
+module.exports = req => `https://${req.hostname}${req.app.locals.SITE_ROOT}/`;